requirements.txt
fastapi
python-dotenv
uvicorn[standard]
openai
crawlee[crawlers]
beautifulsoup4
pydantic==2.11.5
faiss-cpu
numpy
